backend/requirements.txt
asgiref==3.11.0
certifi==2026.2.25
charset-normalizer==3.4.6
Django==6.0
django-cors-headers==4.9.0
django-jazzmin==3.0.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
mysqlclient==2.2.7
PyJWT==2.10.1
PyMySQL==1.1.2
requests==2.33.0
sqlparse==0.5.5
urllib3==2.6.3
